% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{distanceMatrixBetweenFiles}
\alias{distanceMatrixBetweenFiles}
\title{KS Distance between files in FSC/SSC space.}
\usage{
distanceMatrixBetweenFiles(myflowset, cl = NULL)
}
\arguments{
\item{myflowset}{a \code{flowSet}}

\item{cl}{a \code{cluster} object. Default NULL.}
}
\value{
A \code{distance} matrix object.
}
\description{
Computes the ks distance betwen files in a flowset.
}
