% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{findFCSFiles}
\alias{findFCSFiles}
\title{Find fcs files paths from a list matching a gating set}
\usage{
findFCSFiles(gs, files)
}
\arguments{
\item{gs}{a gating set}

\item{files}{a list of FCS files with full paths}
}
\value{
a vector of full paths to files matching the FCS files in the gating set.
}
\description{
Matches based on file name and event number. Assumes the gating set has event number appended to
file name in the sampleName slot.
}
